Added Quick grade endpoint + things
The quick grade endpoint is in the new GradeController. It, like PostSubmission method from SubmissionController, heavely uses the Grader helper. Also, it does not require the user to be authenticated.
Made the Grader helper also return an output string property.
Moved CompilationOutput and Grade from Entities to DTOs (as they don't have a DB table).