TCs are now submitted as text files + more changes
Test cases:
The test case upload rework can be seen in TaskController methods PostTask and PutTask. This change was done to ensure faster upload of large test cases from the front-end (as large TCs caused the browser to freeze when pasted into textareas). Also, the request data limit has been increased to make it possible to upload large TCs.

Test case data now doesn't get loaded whenever the task is loaded. The new TestCaseController handles the loading of individual test cases and returns them as text files. This was also done to ensure a smoother user experience.

Other changes:
Bugfix: Changed the process of writing to solution program's standard input to async. Without async writing the API crashed in some cases, such as when the program would be marked as MLE.

Added SubmittedAt field for GradeDto.

An upper limit of 10 seconds has been put on grading solutions. This was done to prevent abuse.

Each line of output is now right trimmed for whitespaces before being compared to expected output. This was done to validate solutions that are practically correct but would otherwise be rejected. For example, putting an extra space after printing each row of a matrix.